+++ packages/@expo/cli/CHANGELOG.md
@@ -17,6 +17,7 @@
 
 - fix typo in CLI error message when `expo-router` is not installed. ([#36608](https://github.com/expo/expo/pull/36608) by [@abraj](https://github.com/abraj))
 - Bump to `@expo/metro@~54.0.0` ([#39800](https://github.com/expo/expo/pull/39800) by [@kitten](https://github.com/kitten))
+- Remove webpack from `customize` command ([#40005](https://github.com/expo/expo/pull/40005) by [@LeonDvlpmnt](https://github.com/LeonDvlpmnt))
 
 ### ⚠️ Notices